fix(types): always use type specifier for "other" in elsa-type-accept
This is necessary to properly resolve cases where "any type" is
matched with "any form" or "any variable".  Since the dispatch picks
by specificity on the first argument, not specifying the second one
always picked for example "const type and anything" over "type and
form" to resolve to "type and type".